requirements.txt
flask==3.0.3
numpy>=1.24.0,<2.0.0
opencv-python-headless>=4.1.1
mediapipe>=0.10.0
python-dotenv
gdown>=4.7.0
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
PyYAML>=5.3.1
scipy>=1.4.1
Pillow>=7.1.2
pandas>=1.1.4
matplotlib>=3.2.2
requests>=2.23.0
tqdm>=4.41.0
protobuf<4.21.3
seaborn>=0.11.0